Fixes bug on companies#register when a user is connected. Change dependencies in debian/control due to new package for rake
git-svn-id: https://www.ekylibre.org/svn/ekylibre/trunk@1832 67a09383-3dfa-4221-8551-890bac9c277c

+++ app/controllers/sessions_controller.rb
@@ -1,7 +1,11 @@
 class SessionsController < ApplicationController
 
   def new
-    reset_session
+    if session[:user_id]
+      reset_session
+      redirect_to :action=>:new
+      return
+    end
     ActiveRecord::SessionStore::Session.delete_all(["updated_at <= ?", Date.today-1.month])
     session[:side] = false
   end
